+++ the_ray_tracer_challenge/src/canvas/test/mod.rs
@@ -1,3 +1,4 @@
 pub mod color;
 pub mod math;
 pub mod model;
+pub mod ppm;
